cursorrules
added all cursorrules files:
- overview
- functions
- architecture
- pseudocode
- roadmap
- changelog
- lessons learned
- nfr-notes
- workflow

These Cursor rules files are all intended to make Cursor a self-aware coding assistant that self-improves as I use it. The self-improvement happens in the workflow rules file and Cursor can update its own workflows and rules in that workflow rules file when I speak to it and instruct it to do so, if I have feedback on how it's working. These rules files augment Cursor with dynamically updated, interactable context for the whole code base and we'll see how this works because this is the first time I've used something like this.